# frontend/lib/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mushroom Farm App API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/fill-profiles": {
      "get": {
        "summary": "Fill Profiles",
        "operationId": "fill_profiles_api_fill_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FillProfileOut"
                  },
                  "type": "array",
                  "title": "Response Fill Profiles Api Fill Profiles Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Fill Profile",
        "operationId": "create_fill_profile_api_fill_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FillProfileCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FillProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mix-lots": {
      "get": {
        "summary": "List Mix Lots",
        "operationId": "list_mix_lots_api_mix_lots_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MixLotOut"
                  },
                  "type": "array",
                  "title": "Response List Mix Lots Api Mix Lots Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Mix Lot",
        "operationId": "create_mix_lot_api_mix_lots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MixLotCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MixLotOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/species": {
      "get": {
        "summary": "List Species",
        "operationId": "list_species_api_species_get",
        "parameters": [
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MushroomSpeciesOut"
                  },
                  "title": "Response List Species Api Species Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Species",
        "operationId": "create_species_api_species_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MushroomSpeciesCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MushroomSpeciesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/species/{species_id}": {
      "patch": {
        "summary": "Update Species",
        "operationId": "update_species_api_species__species_id__patch",
        "parameters": [
          {
            "name": "species_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Species Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MushroomSpeciesUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MushroomSpeciesOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/species-profiles/{code}": {
      "get": {
        "summary": "Get Species Profile",
        "operationId": "get_species_profile_api_species_profiles__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeciesProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/liquid-cultures": {
      "get": {
        "summary": "List Liquid Cultures",
        "operationId": "list_liquid_cultures_api_liquid_cultures_get",
        "parameters": [
          {
            "name": "active_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Active Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LiquidCultureOut"
                  },
                  "title": "Response List Liquid Cultures Api Liquid Cultures Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Liquid Culture",
        "operationId": "create_liquid_culture_api_liquid_cultures_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LiquidCultureCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiquidCultureOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/grain-types": {
      "get": {
        "summary": "List Grain Types",
        "operationId": "list_grain_types_api_grain_types_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GrainTypeOut"
                  },
                  "type": "array",
                  "title": "Response List Grain Types Api Grain Types Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Grain Type",
        "operationId": "create_grain_type_api_grain_types_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrainTypeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrainTypeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/grain-types/{grain_type_id}": {
      "patch": {
        "summary": "Update Grain Type",
        "operationId": "update_grain_type_api_grain_types__grain_type_id__patch",
        "parameters": [
          {
            "name": "grain_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grain Type Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrainTypeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrainTypeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingredients": {
      "get": {
        "summary": "List Ingredients",
        "operationId": "list_ingredients_api_ingredients_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IngredientOut"
                  },
                  "type": "array",
                  "title": "Response List Ingredients Api Ingredients Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Ingredient",
        "operationId": "create_ingredient_api_ingredients_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngredientCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngredientOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/bags/{bag_id}": {
      "get": {
        "summary": "Get Bag",
        "operationId": "get_bag_api_bags__bag_id__get",
        "parameters": [
          {
            "name": "bag_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bag Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BagDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FillProfileOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "target_dry_kg_per_bag": {
            "type": "number",
            "title": "Target Dry Kg Per Bag"
          },
          "target_water_kg_per_bag": {
            "type": "number",
            "title": "Target Water Kg Per Bag"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "fill_profile_id": {
            "type": "integer",
            "title": "Fill Profile Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "target_dry_kg_per_bag",
          "target_water_kg_per_bag",
          "fill_profile_id"
        ],
        "title": "FillProfileOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}